requirements.txt
# PyTorch CPU wheels index (needed for sentence-transformers on servers without GPU)
--extra-index-url https://download.pytorch.org/whl/cpu

# FastAPI Migration Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0
httpx[http2]==0.25.2
h2==4.2.0
python-multipart==0.0.6
slowapi==0.1.9
pydantic-settings==2.1.0
redis==5.0.1
psutil==5.9.8
python-dotenv==1.0.0

# Existing dependencies
yfinance==0.2.63
scikit-learn==1.4.2
numpy==1.26.4
pandas==2.2.1
transformers==4.38.2
datasets==2.18.0
beautifulsoup4==4.12.3
transformers-interpret==0.10.0
vaderSentiment==3.3.2
requests==2.31.0

# FutureQuant Trader Dependencies
sqlalchemy==2.0.23
sqlmodel==0.0.14
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0
mlflow==2.8.1
pytorch-lightning==2.1.3
ta-lib==0.4.28
polars==0.20.3
celery==5.3.4
apscheduler==3.10.4
minio==7.2.0
boto3==1.34.34

# Development dependencies
pytest==7.4.3
pytest-asyncio==0.21.1 

# RAG / LangChain
langchain==0.2.11
langchain-community==0.2.10
faiss-cpu==1.8.0.post1
sentence-transformers==2.7.0
langchain-text-splitters==0.2.2
torch==2.3.1
pdfminer.six==20231228